refactor: split ApexLogParser into multiple files
- ApexLogParser.ts - Main parser logic
- LogEvents.ts - All log event classes
- LogLineMapping.ts - Mapping of log event types to classes
- types.ts - Types and interfaces
