+++ log-viewer/modules/components/calltree-view/CalltreeView.ts
@@ -27,7 +27,8 @@ import { progressFormatter } from '../../datagrid/format/Progress.js';
 import { RowKeyboardNavigation } from '../../datagrid/module/RowKeyboardNavigation.js';
 import { RowNavigation } from '../../datagrid/module/RowNavigation.js';
 import dataGridStyles from '../../datagrid/style/DataGrid.scss';
-import { ApexLog, LogEvent, type LogEventType } from '../../parsers/ApexLogParser.js';
+import type { ApexLog, LogEvent } from '../../parsers/LogEvents.js';
+import type { LogEventType } from '../../parsers/types.js';
 import { vscodeMessenger } from '../../services/VSCodeExtensionMessenger.js';
 import { globalStyles } from '../../styles/global.styles.js';
 import { isVisible } from '../../Util.js';
