+++ log-viewer/modules/components/calltree-view/module/MiddleRowFocus.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Certinia Inc. All rights reserved.
  */
 import { Module, type RowComponent, type Tabulator } from 'tabulator-tables';
-import type { LogEvent } from '../../../parsers/ApexLogParser';
+import type { LogEvent } from '../../../parsers/LogEvents.js';
 
 type TimedNodeProp = { originalData: LogEvent };
