+++ log-viewer/modules/components/database-view/DatabaseSection.ts
@@ -4,7 +4,7 @@
 import { LitElement, css, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { LogEvent } from '../../parsers/ApexLogParser.js';
+import type { LogEvent } from '../../parsers/LogEvents.js';
 import { globalStyles } from '../../styles/global.styles.js';
 import '../BadgeBase.js';
