+++ log-viewer/modules/parsers/LogLineMapping.ts
@@ -1,0 +1,479 @@
+/*
+ * Copyright (c) 2020 Certinia Inc. All rights reserved.
+ */
+
+import type { ApexLogParser } from './ApexLogParser.js';
+import {
+  BasicExitLine,
+  BasicLogLine,
+  BulkDMLEntry,
+  BulkHeapAllocateLine,
+  CalloutRequestLine,
+  CalloutResponseLine,
+  CodeUnitFinishedLine,
+  CodeUnitStartedLine,
+  ConstructorEntryLine,
+  ConstructorExitLine,
+  CumulativeLimitUsageLine,
+  CumulativeProfilingBeginLine,
+  CumulativeProfilingLine,
+  DMLBeginLine,
+  DMLEndLine,
+  DuplicateDetectionBegin,
+  DuplicateDetectionDetails,
+  DuplicateDetectionRule,
+  DuplicateDetectionSummary,
+  EmailQueueLine,
+  EnteringManagedPackageLine,
+  EventSericePubBeginLine,
+  EventSericePubDetailLine,
+  EventSericePubEndLine,
+  EventSericeSubBeginLine,
+  EventSericeSubDetailLine,
+  EventSericeSubEndLine,
+  ExceptionThrownLine,
+  ExecutionStartedLine,
+  FatalErrorLine,
+  FlowActionCallDetailLine,
+  FlowAssignmentDetailLine,
+  FlowBulkElementBeginLine,
+  FlowBulkElementDetailLine,
+  FlowBulkElementLimitUsageLine,
+  FlowBulkElementNotSupportedLine,
+  FlowCreateInterviewErrorLine,
+  FlowElementAssignmentLine,
+  FlowElementBeginLine,
+  FlowElementDeferredLine,
+  FlowElementErrorLine,
+  FlowElementFaultLine,
+  FlowElementLimitUsageLine,
+  FlowInterviewFinishedLimitUsageLine,
+  FlowInterviewFinishedLine,
+  FlowInterviewPausedLine,
+  FlowInterviewResumedLine,
+  FlowLoopDetailLine,
+  FlowRuleDetailLine,
+  FlowStartInterviewBeginLine,
+  FlowStartInterviewLimitUsageLine,
+  FlowStartInterviewsBeginLine,
+  FlowStartInterviewsErrorLine,
+  FlowStartScheduledRecordsLine,
+  FlowSubflowDetailLine,
+  FlowWaitEventResumingDetailLine,
+  FlowWaitEventWaitingDetailLine,
+  FlowWaitResumingDetailLine,
+  FlowWaitWaitingDetailLine,
+  HeapAllocateLine,
+  HeapDeallocateLine,
+  IdeasQueryExecuteLine,
+  LimitUsageForNSLine,
+  LimitUsageLine,
+  LogEvent,
+  MatchEngineBegin,
+  MethodEntryLine,
+  MethodExitLine,
+  NamedCredentialRequestLine,
+  NamedCredentialResponseDetailLine,
+  NamedCredentialResponseLine,
+  NBANodeBegin,
+  NBANodeDetail,
+  NBANodeEnd,
+  NBANodeError,
+  NBAOfferInvalid,
+  NBAStrategyBegin,
+  NBAStrategyEnd,
+  NBAStrategyError,
+  OrgCacheGetBegin,
+  OrgCachePutBegin,
+  OrgCacheRemoveBegin,
+  PNInvalidAppLine,
+  PNInvalidCertificateLine,
+  PNInvalidNotificationLine,
+  PNNoDevicesLine,
+  PNSentLine,
+  PopTraceFlagsLine,
+  PushTraceFlagsLine,
+  QueryMoreBeginLine,
+  QueryMoreEndLine,
+  QueryMoreIterationsLine,
+  SavepointRollbackLine,
+  SavePointSetLine,
+  SessionCacheGetBegin,
+  SessionCachePutBegin,
+  SessionCacheRemoveBegin,
+  SLAEndLine,
+  SLAEvalMilestoneLine,
+  SLAProcessCaseLine,
+  SOQLExecuteBeginLine,
+  SOQLExecuteEndLine,
+  SOQLExecuteExplainLine,
+  SOSLExecuteBeginLine,
+  SOSLExecuteEndLine,
+  StackFrameVariableListLine,
+  StatementExecuteLine,
+  StaticVariableListLine,
+  SystemConstructorEntryLine,
+  SystemConstructorExitLine,
+  SystemMethodEntryLine,
+  SystemMethodExitLine,
+  SystemModeEnterLine,
+  SystemModeExitLine,
+  TestingLimitsLine,
+  TotalEmailRecipientsQueuedLine,
+  UserDebugLine,
+  UserInfoLine,
+  ValidationErrorLine,
+  ValidationFormulaLine,
+  ValidationPassLine,
+  ValidationRuleLine,
+  VariableAssignmentLine,
+  VariableScopeBeginLine,
+  VFApexCallEndLine,
+  VFApexCallStartLine,
+  VFDeserializeContinuationStateBegin,
+  VFDeserializeViewstateBeginLine,
+  VFFormulaEndLine,
+  VFFormulaStartLine,
+  VFPageMessageLine,
+  VFSeralizeViewStateStartLine,
+  VFSerializeContinuationStateBegin,
+  WFActionLine,
+  WFActionsEndLine,
+  WFActionTaskLine,
+  WFApprovalLine,
+  WFApprovalRemoveLine,
+  WFApprovalSubmitLine,
+  WFApprovalSubmitterLine,
+  WFAssignLine,
+  WFCriteriaBeginLine,
+  WFEmailAlertLine,
+  WFEmailSentLine,
+  WFEnqueueActionsLine,
+  WFEscalationActionLine,
+  WFEvalEntryCriteriaLine,
+  WFFieldUpdateLine,
+  WFFlowActionDetailLine,
+  WFFlowActionErrorDetailLine,
+  WFFlowActionErrorLine,
+  WFFormulaLine,
+  WFNextApproverLine,
+  WFOutboundMsgLine,
+  WFProcessFoundLine,
+  WFProcessNode,
+  WFReassignRecordLine,
+  WFResponseNotifyLine,
+  WFRuleEntryOrderLine,
+  WFRuleEvalBeginLine,
+  WFRuleEvalValueLine,
+  WFRuleFilterLine,
+  WFRuleInvocationLine,
+  WFSoftRejectLine,
+  WFSpoolActionBeginLine,
+  WFTimeTriggerLine,
+  XDSDetailLine,
+  XDSResponseDetailLine,
+  XDSResponseErrorLine,
+  XDSResponseLine,
+} from './LogEvents.js';
+import type { LogEventType, LogLineConstructor } from './types.js';
+
+export function getLogEventClass(
+  eventName: LogEventType,
+): LogLineConstructor<ApexLogParser, LogEvent> | null | undefined {
+  if (!eventName) {
+    return null;
+  }
+
+  // Fast path for the most commonly occuring types
+  switch (eventName) {
+    case 'METHOD_ENTRY':
+      return MethodEntryLine;
+
+    case 'METHOD_EXIT':
+      return MethodExitLine;
+
+    case 'CONSTRUCTOR_ENTRY':
+      return ConstructorEntryLine;
+
+    case 'CONSTRUCTOR_EXIT':
+      return ConstructorExitLine;
+
+    default:
+      break;
+  }
+
+  // Handle all other types
+  const logType = lineTypeMap.get(eventName);
+  if (logType) {
+    return logType;
+  } else if (basicLogEvents.indexOf(eventName) !== -1) {
+    return BasicLogLine;
+  } else if (basicExitLogEvents.indexOf(eventName) !== -1) {
+    return BasicExitLine;
+  }
+
+  return null;
+}
+
+export const lineTypeMap: ReadonlyMap<
+  LogEventType,
+  LogLineConstructor<ApexLogParser, LogEvent>
+> = new Map<LogEventType, LogLineConstructor<ApexLogParser, LogEvent>>([
+  ['BULK_DML_RETRY', BulkDMLEntry],
+  ['BULK_HEAP_ALLOCATE', BulkHeapAllocateLine],
+  ['CALLOUT_REQUEST', CalloutRequestLine],
+  ['CALLOUT_RESPONSE', CalloutResponseLine],
+  ['NAMED_CREDENTIAL_REQUEST', NamedCredentialRequestLine],
+  ['NAMED_CREDENTIAL_RESPONSE', NamedCredentialResponseLine],
+  ['NAMED_CREDENTIAL_RESPONSE_DETAIL', NamedCredentialResponseDetailLine],
+  ['CONSTRUCTOR_ENTRY', ConstructorEntryLine],
+  ['CONSTRUCTOR_EXIT', ConstructorExitLine],
+  ['EMAIL_QUEUE', EmailQueueLine],
+  ['METHOD_ENTRY', MethodEntryLine],
+  ['METHOD_EXIT', MethodExitLine],
+  ['SYSTEM_CONSTRUCTOR_ENTRY', SystemConstructorEntryLine],
+  ['SYSTEM_CONSTRUCTOR_EXIT', SystemConstructorExitLine],
+  ['SYSTEM_METHOD_ENTRY', SystemMethodEntryLine],
+  ['SYSTEM_METHOD_EXIT', SystemMethodExitLine],
+  ['CODE_UNIT_STARTED', CodeUnitStartedLine],
+  ['CODE_UNIT_FINISHED', CodeUnitFinishedLine],
+  ['VF_APEX_CALL_START', VFApexCallStartLine],
+  ['VF_APEX_CALL_END', VFApexCallEndLine],
+  ['VF_DESERIALIZE_VIEWSTATE_BEGIN', VFDeserializeViewstateBeginLine],
+  ['VF_EVALUATE_FORMULA_BEGIN', VFFormulaStartLine],
+  ['VF_EVALUATE_FORMULA_END', VFFormulaEndLine],
+  ['VF_SERIALIZE_CONTINUATION_STATE_BEGIN', VFSerializeContinuationStateBegin],
+  ['VF_DESERIALIZE_CONTINUATION_STATE_BEGIN', VFDeserializeContinuationStateBegin],
+  ['VF_SERIALIZE_VIEWSTATE_BEGIN', VFSeralizeViewStateStartLine],
+  ['VF_PAGE_MESSAGE', VFPageMessageLine],
+  ['DML_BEGIN', DMLBeginLine],
+  ['DML_END', DMLEndLine],
+  ['IDEAS_QUERY_EXECUTE', IdeasQueryExecuteLine],
+  ['SOQL_EXECUTE_BEGIN', SOQLExecuteBeginLine],
+  ['SOQL_EXECUTE_END', SOQLExecuteEndLine],
+  ['SOQL_EXECUTE_EXPLAIN', SOQLExecuteExplainLine],
+  ['SOSL_EXECUTE_BEGIN', SOSLExecuteBeginLine],
+  ['SOSL_EXECUTE_END', SOSLExecuteEndLine],
+  ['HEAP_ALLOCATE', HeapAllocateLine],
+  ['HEAP_DEALLOCATE', HeapDeallocateLine],
+  ['STATEMENT_EXECUTE', StatementExecuteLine],
+  ['VARIABLE_SCOPE_BEGIN', VariableScopeBeginLine],
+  ['VARIABLE_ASSIGNMENT', VariableAssignmentLine],
+  ['USER_INFO', UserInfoLine],
+  ['USER_DEBUG', UserDebugLine],
+  ['CUMULATIVE_LIMIT_USAGE', CumulativeLimitUsageLine],
+  ['CUMULATIVE_PROFILING', CumulativeProfilingLine],
+  ['CUMULATIVE_PROFILING_BEGIN', CumulativeProfilingBeginLine],
+  ['LIMIT_USAGE', LimitUsageLine],
+  ['LIMIT_USAGE_FOR_NS', LimitUsageForNSLine],
+  ['NBA_NODE_BEGIN', NBANodeBegin],
+  ['NBA_NODE_DETAIL', NBANodeDetail],
+  ['NBA_NODE_END', NBANodeEnd],
+  ['NBA_NODE_ERROR', NBANodeError],
+  ['NBA_OFFER_INVALID', NBAOfferInvalid],
+  ['NBA_STRATEGY_BEGIN', NBAStrategyBegin],
+  ['NBA_STRATEGY_END', NBAStrategyEnd],
+  ['NBA_STRATEGY_ERROR', NBAStrategyError],
+  ['POP_TRACE_FLAGS', PopTraceFlagsLine],
+  ['PUSH_TRACE_FLAGS', PushTraceFlagsLine],
+  ['QUERY_MORE_BEGIN', QueryMoreBeginLine],
+  ['QUERY_MORE_END', QueryMoreEndLine],
+  ['QUERY_MORE_ITERATIONS', QueryMoreIterationsLine],
+  ['TOTAL_EMAIL_RECIPIENTS_QUEUED', TotalEmailRecipientsQueuedLine],
+  ['SAVEPOINT_ROLLBACK', SavepointRollbackLine],
+  ['SAVEPOINT_SET', SavePointSetLine],
+  ['STACK_FRAME_VARIABLE_LIST', StackFrameVariableListLine],
+  ['STATIC_VARIABLE_LIST', StaticVariableListLine],
+  ['SYSTEM_MODE_ENTER', SystemModeEnterLine],
+  ['SYSTEM_MODE_EXIT', SystemModeExitLine],
+  ['EXECUTION_STARTED', ExecutionStartedLine],
+  ['ENTERING_MANAGED_PKG', EnteringManagedPackageLine],
+  ['EVENT_SERVICE_PUB_BEGIN', EventSericePubBeginLine],
+  ['EVENT_SERVICE_PUB_END', EventSericePubEndLine],
+  ['EVENT_SERVICE_PUB_DETAIL', EventSericePubDetailLine],
+  ['EVENT_SERVICE_SUB_BEGIN', EventSericeSubBeginLine],
+  ['EVENT_SERVICE_SUB_DETAIL', EventSericeSubDetailLine],
+  ['EVENT_SERVICE_SUB_END', EventSericeSubEndLine],
+  ['FLOW_START_INTERVIEWS_BEGIN', FlowStartInterviewsBeginLine],
+  ['FLOW_START_INTERVIEWS_ERROR', FlowStartInterviewsErrorLine],
+  ['FLOW_START_INTERVIEW_BEGIN', FlowStartInterviewBeginLine],
+  ['FLOW_START_INTERVIEW_LIMIT_USAGE', FlowStartInterviewLimitUsageLine],
+  ['FLOW_START_SCHEDULED_RECORDS', FlowStartScheduledRecordsLine],
+  ['FLOW_CREATE_INTERVIEW_ERROR', FlowCreateInterviewErrorLine],
+  ['FLOW_ELEMENT_BEGIN', FlowElementBeginLine],
+  ['FLOW_ELEMENT_DEFERRED', FlowElementDeferredLine],
+  ['FLOW_ELEMENT_ERROR', FlowElementErrorLine],
+  ['FLOW_ELEMENT_FAULT', FlowElementFaultLine],
+  ['FLOW_ELEMENT_LIMIT_USAGE', FlowElementLimitUsageLine],
+  ['FLOW_INTERVIEW_FINISHED_LIMIT_USAGE', FlowInterviewFinishedLimitUsageLine],
+  ['FLOW_SUBFLOW_DETAIL', FlowSubflowDetailLine],
+  ['FLOW_VALUE_ASSIGNMENT', FlowElementAssignmentLine],
+  ['FLOW_WAIT_EVENT_RESUMING_DETAIL', FlowWaitEventResumingDetailLine],
+  ['FLOW_WAIT_EVENT_WAITING_DETAIL', FlowWaitEventWaitingDetailLine],
+  ['FLOW_WAIT_RESUMING_DETAIL', FlowWaitResumingDetailLine],
+  ['FLOW_WAIT_WAITING_DETAIL', FlowWaitWaitingDetailLine],
+  ['FLOW_INTERVIEW_FINISHED', FlowInterviewFinishedLine],
+  ['FLOW_INTERVIEW_PAUSED', FlowInterviewPausedLine],
+  ['FLOW_INTERVIEW_RESUMED', FlowInterviewResumedLine],
+  ['FLOW_ACTIONCALL_DETAIL', FlowActionCallDetailLine],
+  ['FLOW_ASSIGNMENT_DETAIL', FlowAssignmentDetailLine],
+  ['FLOW_LOOP_DETAIL', FlowLoopDetailLine],
+  ['FLOW_RULE_DETAIL', FlowRuleDetailLine],
+  ['FLOW_BULK_ELEMENT_BEGIN', FlowBulkElementBeginLine],
+  ['FLOW_BULK_ELEMENT_DETAIL', FlowBulkElementDetailLine],
+  ['FLOW_BULK_ELEMENT_LIMIT_USAGE', FlowBulkElementLimitUsageLine],
+  ['FLOW_BULK_ELEMENT_NOT_SUPPORTED', FlowBulkElementNotSupportedLine],
+  ['MATCH_ENGINE_BEGIN', MatchEngineBegin],
+  ['ORG_CACHE_PUT_BEGIN', OrgCachePutBegin],
+  ['ORG_CACHE_GET_BEGIN', OrgCacheGetBegin],
+  ['ORG_CACHE_REMOVE_BEGIN', OrgCacheRemoveBegin],
+  ['PUSH_NOTIFICATION_INVALID_APP', PNInvalidAppLine],
+  ['PUSH_NOTIFICATION_INVALID_CERTIFICATE', PNInvalidCertificateLine],
+  ['PUSH_NOTIFICATION_INVALID_NOTIFICATION', PNInvalidNotificationLine],
+  ['PUSH_NOTIFICATION_NO_DEVICES', PNNoDevicesLine],
+  ['PUSH_NOTIFICATION_SENT', PNSentLine],
+  ['SESSION_CACHE_PUT_BEGIN', SessionCachePutBegin],
+  ['SESSION_CACHE_GET_BEGIN', SessionCacheGetBegin],
+  ['SESSION_CACHE_REMOVE_BEGIN', SessionCacheRemoveBegin],
+  ['SLA_END', SLAEndLine],
+  ['SLA_EVAL_MILESTONE', SLAEvalMilestoneLine],
+  ['SLA_PROCESS_CASE', SLAProcessCaseLine],
+  ['TESTING_LIMITS', TestingLimitsLine],
+  ['VALIDATION_ERROR', ValidationErrorLine],
+  ['VALIDATION_FORMULA', ValidationFormulaLine],
+  ['VALIDATION_PASS', ValidationPassLine],
+  ['VALIDATION_RULE', ValidationRuleLine],
+  ['WF_FLOW_ACTION_ERROR', WFFlowActionErrorLine],
+  ['WF_FLOW_ACTION_ERROR_DETAIL', WFFlowActionErrorDetailLine],
+  ['WF_FIELD_UPDATE', WFFieldUpdateLine],
+  ['WF_RULE_EVAL_BEGIN', WFRuleEvalBeginLine],
+  ['WF_RULE_EVAL_VALUE', WFRuleEvalValueLine],
+  ['WF_RULE_FILTER', WFRuleFilterLine],
+  ['WF_CRITERIA_BEGIN', WFCriteriaBeginLine],
+  ['WF_FORMULA', WFFormulaLine],
+  ['WF_ACTION', WFActionLine],
+  ['WF_ACTIONS_END', WFActionsEndLine],
+  ['WF_ACTION_TASK', WFActionTaskLine],
+  ['WF_APPROVAL', WFApprovalLine],
+  ['WF_APPROVAL_REMOVE', WFApprovalRemoveLine],
+  ['WF_APPROVAL_SUBMIT', WFApprovalSubmitLine],
+  ['WF_APPROVAL_SUBMITTER', WFApprovalSubmitterLine],
+  ['WF_ASSIGN', WFAssignLine],
+  ['WF_EMAIL_ALERT', WFEmailAlertLine],
+  ['WF_EMAIL_SENT', WFEmailSentLine],
+  ['WF_ENQUEUE_ACTIONS', WFEnqueueActionsLine],
+  ['WF_ESCALATION_ACTION', WFEscalationActionLine],
+  ['WF_EVAL_ENTRY_CRITERIA', WFEvalEntryCriteriaLine],
+  ['WF_FLOW_ACTION_DETAIL', WFFlowActionDetailLine],
+  ['WF_NEXT_APPROVER', WFNextApproverLine],
+  ['WF_OUTBOUND_MSG', WFOutboundMsgLine],
+  ['WF_PROCESS_FOUND', WFProcessFoundLine],
+  ['WF_PROCESS_NODE', WFProcessNode],
+  ['WF_REASSIGN_RECORD', WFReassignRecordLine],
+  ['WF_RESPONSE_NOTIFY', WFResponseNotifyLine],
+  ['WF_RULE_ENTRY_ORDER', WFRuleEntryOrderLine],
+  ['WF_RULE_INVOCATION', WFRuleInvocationLine],
+  ['WF_SOFT_REJECT', WFSoftRejectLine],
+  ['WF_SPOOL_ACTION_BEGIN', WFSpoolActionBeginLine],
+  ['WF_TIME_TRIGGER', WFTimeTriggerLine],
+  ['EXCEPTION_THROWN', ExceptionThrownLine],
+  ['FATAL_ERROR', FatalErrorLine],
+  ['XDS_DETAIL', XDSDetailLine],
+  ['XDS_RESPONSE', XDSResponseLine],
+  ['XDS_RESPONSE_DETAIL', XDSResponseDetailLine],
+  ['XDS_RESPONSE_ERROR', XDSResponseErrorLine],
+  ['DUPLICATE_DETECTION_BEGIN', DuplicateDetectionBegin],
+  ['DUPLICATE_DETECTION_RULE_INVOCATION', DuplicateDetectionRule],
+  ['DUPLICATE_DETECTION_MATCH_INVOCATION_DETAILS', DuplicateDetectionDetails],
+  ['DUPLICATE_DETECTION_MATCH_INVOCATION_SUMMARY', DuplicateDetectionSummary],
+]);
+
+const basicLogEvents: LogEventType[] = [
+  'BULK_COUNTABLE_STATEMENT_EXECUTE',
+  'TEMPLATE_PROCESSING_ERROR',
+  'EXTERNAL_SERVICE_REQUEST',
+  'FLOW_CREATE_INTERVIEW_BEGIN',
+  'FLOW_CREATE_INTERVIEW_END',
+  'VARIABLE_SCOPE_END',
+  'PUSH_NOTIFICATION_NOT_ENABLED',
+  'SLA_NULL_START_DATE',
+  'TEMPLATE_PROCESSING_ERROR',
+  'VALIDATION_FAIL',
+  `WF_FLOW_ACTION_BEGIN`,
+  'WF_FLOW_ACTION_END',
+  'WF_ESCALATION_RULE',
+  'WF_HARD_REJECT',
+  'WF_NO_PROCESS_FOUND',
+  'WF_TIME_TRIGGERS_BEGIN',
+  'WF_KNOWLEDGE_ACTION',
+  'WF_SEND_ACTION',
+  'WAVE_APP_LIFECYCLE',
+  'WF_QUICK_CREATE',
+  'WF_APEX_ACTION',
+  'INVOCABLE_ACTION_DETAIL',
+  'INVOCABLE_ACTION_ERROR',
+  'FLOW_COLLECTION_PROCESSOR_DETAIL',
+  'FLOW_SCHEDULED_PATH_QUEUED',
+  'ROUTE_WORK_ACTION',
+  'ADD_SKILL_REQUIREMENT_ACTION',
+  'ADD_SCREEN_POP_ACTION',
+  'CALLOUT_REQUEST_PREPARE',
+  'CALLOUT_REQUEST_FINALIZE',
+  'FUNCTION_INVOCATION_REQUEST',
+  'APP_CONTAINER_INITIATED',
+  'FUNCTION_INVOCATION_RESPONSE',
+  'XDS_REQUEST_DETAIL',
+  'EXTERNAL_SERVICE_RESPONSE',
+  'DATAWEAVE_USER_DEBUG',
+  'USER_DEBUG_FINER',
+  'USER_DEBUG_FINEST',
+  'USER_DEBUG_FINE',
+  'USER_DEBUG_DEBUG',
+  'USER_DEBUG_INFO',
+  'USER_DEBUG_WARN',
+  'USER_DEBUG_ERROR',
+  'VF_APEX_CALL',
+  'HEAP_DUMP',
+  'SCRIPT_EXECUTION',
+  'SESSION_CACHE_MEMORY_USAGE',
+  'ORG_CACHE_MEMORY_USAGE',
+  'AE_PERSIST_VALIDATION',
+  'REFERENCED_OBJECT_LIST',
+  'DUPLICATE_RULE_FILTER',
+  'DUPLICATE_RULE_FILTER_RESULT',
+  'DUPLICATE_RULE_FILTER_VALUE',
+  'TEMPLATED_ASSET',
+  'TRANSFORMATION_SUMMARY',
+  'RULES_EXECUTION_SUMMARY',
+  'ASSET_DIFF_SUMMARY',
+  'ASSET_DIFF_DETAIL',
+  'RULES_EXECUTION_DETAIL',
+  'JSON_DIFF_SUMMARY',
+  'JSON_DIFF_DETAIL',
+  'MATCH_ENGINE_INVOCATION',
+];
+
+const basicExitLogEvents: LogEventType[] = [
+  'FLOW_START_INTERVIEW_END',
+  'VF_DESERIALIZE_VIEWSTATE_END',
+  'VF_SERIALIZE_VIEWSTATE_END',
+  'CUMULATIVE_LIMIT_USAGE_END',
+  'CUMULATIVE_PROFILING_END',
+  'EXECUTION_FINISHED',
+  'FLOW_START_INTERVIEWS_END',
+  'FLOW_ELEMENT_END',
+  'FLOW_BULK_ELEMENT_END',
+  'WF_RULE_EVAL_END',
+  'WF_RULE_NOT_EVALUATED',
+  'WF_CRITERIA_END',
+  'DUPLICATE_DETECTION_END',
+  'VF_SERIALIZE_CONTINUATION_STATE_END',
+  'VF_DESERIALIZE_CONTINUATION_STATE_END',
+  'MATCH_ENGINE_END',
+  'ORG_CACHE_PUT_END',
+  'ORG_CACHE_GET_END',
+  'ORG_CACHE_REMOVE_END',
+  'SESSION_CACHE_PUT_END',
+  'SESSION_CACHE_GET_END',
+  'SESSION_CACHE_REMOVE_END',
+];
